refactor(SkillBreakdown): Refactor duration handling to use TimeSpan and ticks
Standardize duration representation across statistics and view models by replacing ulong millisecond durations with TimeSpan or tick-based values. Centralize aggregation and average calculation logic in DataStatisticsExtensions, and update DataStatistics to use double for average values. Update all related models and view models to use DurationTicks, improving precision, reducing code duplication, and enhancing maintainability.

+++ StarResonanceDpsAnalysis.WPF/ViewModels/DataStatistics.cs
@@ -7,7 +7,7 @@ namespace StarResonanceDpsAnalysis.WPF.ViewModels;
 /// </summary>
 public partial class DataStatistics : BaseViewModel
 {
-    [ObservableProperty] private long _average;
+    [ObservableProperty] private double _average;
     [ObservableProperty] private double _critRate;
     [ObservableProperty] private int _critCount;
     [ObservableProperty] private int _luckyCount;
